- Logic and API for models

+++ HW5.Contracts/Request/CreateClientRequest.cs
@@ -8,6 +8,6 @@ namespace HW5.Contracts.Request
         public string LastName { get; set; }
         public string MiddleName { get; set; }
         public DateTime BirthDate { get; set; }
-        public uint PhoneNumber { get; set; }
+        public ulong PhoneNumber { get; set; }
     }
 }
